mpl: pair<int,int> and mpl::Point change to odb::Point, removal of leftover floats

+++ src/mpl/src/mpl-util.h
@@ -210,7 +210,7 @@ inline int64_t computeDistToNearestRegion(
     }
   }
 
-  return static_cast<int64_t>(std::sqrt(smallest_distance));
+  return std::sqrt(smallest_distance);
 }
 
 }  // namespace mpl
